pyproject.toml
[project]
name = "django-mini-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.2",
    "django-extensions==4.1",
    "django-filter==25.1",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.0",
    "drf-yasg==1.21.10",
    "ipython==9.3.0",
    "psycopg2==2.9.10",
]

[dependency-groups]
dev = [
    "black==25.1.0",
    "isort==6.0.1",
    "mypy==1.16.0",
    "ruff==0.11.12",
]
